refactor(template-compiler): extract shared host_codegen module
