+++ crates/template-compiler/src/lib.rs
@@ -10,6 +10,7 @@ mod codegen;
 mod directive_codegen;
 mod extract;
 mod factory_codegen;
+pub mod host_codegen;
 pub mod i18n;
 mod injectable_codegen;
 mod ng_module_codegen;
